+++ source/direct3d9/Query.cpp
@@ -231,7 +231,7 @@ namespace Direct3D9
 		else
 		{
 			T data;
-			hr = InternalPointer->GetData( &data, sizeof(T), flags );
+			hr = InternalPointer->GetData( &data, (DWORD) sizeof(T), flags );
 			RECORD_D3D9( hr );
 			return data;
 		}
